+++ CHANGELOG.md
@@ -10,3 +10,9 @@
   - Migrated `logstash-input-snmptrap` client from ruby-snmp to SNMP4J
     - The event `message` content was changed from the ruby `SNMP::SNMPv1_Trap` object `inspect` representation
       to a hash dump based on the RFC definition. It also adds the fields as `[@metadata][input][snmptrap][pdu][<name>]` metadata.
+  - Standardized the `logstash-input-snmp` and `logstash-input-snmptrap` settings, supported MIB formats and features.
+    - Both plugins now support .dic (`libsmi`) and .yaml (`ruby-snmp`) MIB formats.
+    - Added the `oid_mapping_format` option to configure how the OID field is mapped in the Logstash event.
+    - The `logstash-input-snmptrap` now supports the `logstash-input-snmp` features: `mib_paths`, `use_provided_mibs`, `oid_root_skip`, `oid_path_length`.
+  - The `logstash-input-snmptrap` `yamlmibdir` was deprecated in favor of `mib_paths`.
+  - Updated the `logstash-input-snmptrap` docs to include the same `logstash-input-snmp` instructions on how importing MIB files.